% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{pgm_auth_renv}
\alias{pgm_auth_renv}
\title{Permite autenticar via pacote 'gmailr' a partir de arquivo 'Renviron'}
\usage{
pgm_auth_renv(
  oauth_name = "GOOGLE_OAUTH_NAME",
  oauth_client_id = "GOOGLE_OAUTH_CLIENT_ID",
  oauth_client_secret_key = "GOOGLE_OAUTH_CLIENT_SECRET_KEY",
  token = "GOOGLE_TOKEN",
  email = "GOOGLE_EMAIL",
  cache = "GOOGLE_CACHE",
  use_oob = FALSE
)
}
\arguments{
\item{oauth_name}{Padrão 'GOOGLE_OAUTH_NAME'}

\item{oauth_client_id}{Padrão 'GOOGLE_OAUTH_CLIENT_ID'}

\item{oauth_client_secret_key}{Padrão 'GOOGLE_OAUTH_CLIENT_SECRET_KEY'}

\item{token}{Padrão 'GOOGLE_TOKEN'}

\item{email}{Padrão 'GOOGLE_EMAIL'}

\item{cache}{Padrão 'GOOGLE_CACHE'}

\item{use_oob}{Padrão 'FALSE'}
}
\value{
Autenticação via 'gmailr' à API do 'gmail' a partir de arquivo 'Renviron'
}
\description{
Permite autenticar via pacote 'gmailr' a partir de arquivo 'Renviron'
}
