% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{pgm_labels}
\alias{pgm_labels}
\title{Cria 'data.frame' com a relação de marcadores associados ao e-mail com o respectivo
identificador 'id' e tipo 'type'}
\usage{
pgm_labels(label, type, ...)
}
\arguments{
\item{label}{Marcador do gmail para filtro. Se ausente, retorna todos os marcadores}

\item{type}{Tipo de marcador para filtro. Se ausente, retorna todos os tipos.
Tipos permitidos: 'system' para marcadores de sistema e 'user' para marcadores
de usuários}

\item{...}{Argumentos a serem repassados à função 'gm_labels' do pacote 'gmailr'}
}
\value{
'data.frame' com a relação de marcadores
}
\description{
Cria 'data.frame' com a relação de marcadores associados ao e-mail com o respectivo
identificador 'id' e tipo 'type'
}
