% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{pgm_listmessages}
\alias{pgm_listmessages}
\title{Relaciona e-mails com os respectivos dados a partir de um marcador}
\usage{
pgm_listmessages(label, num_results = 1, num_results_by = "thread")
}
\arguments{
\item{label}{Marcador do gmail para filtro. Campo obrigatório.}

\item{num_results}{Número de resultados de e-mails a partir do filtro 'label'}

\item{num_results_by}{Número de resultados de e-mails a partir de e-mails
e conversas 'all', apenas conversas 'thread' ou apenas e-mails 'message'}
}
\description{
Relaciona e-mails com os respectivos dados a partir de um marcador
}
