% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{pgm_messages}
\alias{pgm_messages}
\title{Cria 'data.frame' com e-mails a partir de parâmetros repassados à função 
'gm_messages'}
\usage{
pgm_messages(...)
}
\arguments{
\item{...}{Argumentos a serem repassados à função 'gm_messages' do pacote 
'gmailr'. Se ausente, retorna os 100 mais recentes e-mails, independente da
vinculação a determinada conversa}
}
\value{
'data.frame' com a relação de e-mails
}
\description{
Cria 'data.frame' com e-mails a partir de parâmetros repassados à função 
'gm_messages'
}
