% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{pgm_read}
\alias{pgm_read}
\title{Permite a leitura de dados de e-mails a partir de conversas}
\usage{
pgm_read(thread_id, message_id)
}
\arguments{
\item{thread_id}{Identificador de conversa 'threads.id'}

\item{message_id}{Identificador de e-mail 'messages.threadId' vinculado a
determinada conversa}
}
\value{
'data.frame' com dados de e-mails vinculados a determinada conversa
}
\description{
Permite a leitura de dados de e-mails a partir de conversas
}
