% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablehtml.R
\name{pgm_tablehtml}
\alias{pgm_tablehtml}
\title{Tabela padrão para envio no corpo do e-mail a ser criado a partir do pacote
'ppgmailr'}
\usage{
pgm_tablehtml(
  obj,
  width = NULL,
  bg_title = "#374049",
  color_title = "white",
  bg_column = "#e2e6e9",
  ...
)
}
\arguments{
\item{obj}{Objeto a ser convertido em tabela 'html'}

\item{width}{Vetor com o tamanho de cada coluna da tabela em 'pixels'. Valor
ausente 'NULL' por padrão}

\item{bg_title}{Cor de fundo do título. Valor padrão '#374049'}

\item{color_title}{Cor do texto. Valor padrão 'white' ou '#ffffff'}

\item{bg_column}{Cor de fundo da primeira coluna da tabela. Valor padrão #e2e6e9'}

\item{...}{Outros parâmetros para a função 'tableHTML'}
}
\value{
Tabela em html a ser usada no corpo do e-mail
}
\description{
Tabela padrão para envio no corpo do e-mail a ser criado a partir do pacote
'ppgmailr'
}
