% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{pgm_threads}
\alias{pgm_threads}
\title{Cria 'data.frame' com conversas do gmail (e-mails agrupados por temas) a partir
de parâmetros repassados à função 'gm_threads'}
\usage{
pgm_threads(...)
}
\arguments{
\item{...}{Argumentos a serem repassados à função 'gm_threads' do pacote 'gmailr'.
Se ausentes, retorna as 100 mais recentes conversas}
}
\value{
'data.frame' com a relação de conversas
}
\description{
Cria 'data.frame' com conversas do gmail (e-mails agrupados por temas) a partir
de parâmetros repassados à função 'gm_threads'
}
