% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{pgm_write}
\alias{pgm_write}
\title{Permite escrever e-mails para envio via API do 'gmail'}
\usage{
pgm_write(
  from,
  to,
  subject = NULL,
  body = NULL,
  file = NULL,
  send = FALSE,
  thread_id = NULL
)
}
\arguments{
\item{from}{Endereço de e-mail do 'gmail' válido. Corresponde ao campo 'De'.}

\item{to}{Endereço de e-mail do 'gmail' válido. Corresponde ao campo 'Para'.}

\item{subject}{Assunto do e-mail. Padrão ausente 'NULL'}

\item{body}{Corpo do e-mail. Padrão ausente 'NULL'}

\item{file}{Arquivo de anexo. Padrão ausente 'NULL'}

\item{send}{Se 'TRUE' encaminha o e-mail a partir da função}

\item{thread_id}{Identificador de conversa a ser vinculado ao e-mail. Padrão
ausente 'NULL'}
}
\value{
'mime' de e-mail válido para envio ou 'id' de conversa, caso o e-mail 
seja enviado a partir da própria função ('send' igual 'TRUE')
}
\description{
Permite escrever e-mails para envio via API do 'gmail'
}
